Clang: float -> qreal where it matches Qt API

+++ src/libcalamaresui/utils/ImageRegistry.cpp
@@ -53,14 +53,14 @@ ImageRegistry::icon( const QString& image, CalamaresUtils::ImageMode mode )
 
 
 qint64
-ImageRegistry::cacheKey( const QSize& size, float opacity, QColor tint )
+ImageRegistry::cacheKey( const QSize& size, qreal opacity, QColor tint )
 {
     return size.width() * 100 + size.height() * 10 + int( opacity * 100.0 ) + tint.value();
 }
 
 
 QPixmap
-ImageRegistry::pixmap( const QString& image, const QSize& size, CalamaresUtils::ImageMode mode, float opacity, QColor tint )
+ImageRegistry::pixmap( const QString& image, const QSize& size, CalamaresUtils::ImageMode mode, qreal opacity, QColor tint )
 {
     QHash< qint64, QPixmap > subsubcache;
     QHash< int, QHash< qint64, QPixmap > > subcache;
@@ -136,7 +136,7 @@ ImageRegistry::pixmap( const QString& image, const QSize& size, CalamaresUtils::
 
 
 void
-ImageRegistry::putInCache( const QString& image, const QSize& size, CalamaresUtils::ImageMode mode, float opacity, const QPixmap& pixmap, QColor tint )
+ImageRegistry::putInCache( const QString& image, const QSize& size, CalamaresUtils::ImageMode mode, qreal opacity, const QPixmap& pixmap, QColor tint )
 {
 //    cDebug( LOGVERBOSE ) << Q_FUNC_INFO << "Adding to image cache:" << image << size << mode;
